refactor(DynamicHyperlink): PDTVT-685 Require the onFetch parameter to handle retrieving url data
