+++ packages/DynamicHyperlink/package.json
@@ -16,7 +16,7 @@
   },
   "dependencies": {
     "@babel/runtime-corejs2": "^7.3.1",
-    "@paprika/l10n": "^1.1.2",
+    "@paprika/l10n": "^1.1.3-alpha.0",
     "@paprika/stylers": "^1.0.0",
     "@paprika/tabs": "^1.1.0",
     "@paprika/tokens": "^1.0.0",
